refactor: move RNN VAE model to model/ directory
